// libs/sentry/src/sentry.module.interface.ts
export type SentryFilterFunction = (exception: unknown) => boolean;
export interface SentryInterceptorOptionsFilter {
  filter?: SentryFilterFunction;
  type: any;
}
export interface SentryModuleOptions {
  debug?: boolean;
  dsn?: string;
  environment?: string;
  logLevel?: LogLevel;
  release?: string;
  context?: "Http" | "Ws" | "Rpc" | "GraphQL";
  extra?: Record<string, unknown>;
  filters?: SentryInterceptorOptionsFilter[];
  fingerprint?: string[];
  level?: Severity;
  tags?: Record<string, string>;
  // https://github.com/getsentry/sentry-javascript/blob/master/packages/node/src/handlers.ts#L163
  request?: boolean;
  serverName?: boolean;
  // https://github.com/getsentry/sentry-javascript/blob/master/packages/node/src/handlers.ts#L16
  transaction?: boolean | "path" | "methodPath" | "handler";
  user?: boolean | string[];
  version?: boolean;
}